Fix to dagnet execution & dependency pruning
Summary:
Matt uyt reported (1) a very infrequent assertion failure at net.cc worker function. This was caused because an operator, that was not a chain, was scheduled in the job queue. This was possibly to happen since our DAGnet operator graph is graph of operators, and not chains. The dependency prunign that I introduced last week exposed this problem since it removed some "middle-to-chain" dependencies when computing the chains. (It is bit hard to explain).

This diff attempts to fix the problem by only allowing scheduling of chains. I added, in addition, extra check to confirm that all parents of all nodes were indeed executed before starting next roud. This adds additional safety and breakpoint to see if there are still problems.

I also fixed a bug in the operator graph pruning that made pruning less effective.

(1) Matt's report:
https://www.prod.facebook.com/groups/1405155842844877/permalink/1639428779417581/

Reviewed By: dzhulgakov

Differential Revision: D4531424

fbshipit-source-id: 80fa7def6e8aff6910ebf0d9d5fef15ff20e0aec