Provide ability to specify more types for ConstantFillOp
Summary:
It looks like for most of the types there is not way we can get them (except
the results of operation on top of some other tensor), that was pretty
unfortunate for cases when we want to do partial type inference (I was trying
to do so in D4408771).

This diff is adding more possible types for ConstantFillOp. Please let me know
if I'm missing anything. The only part that worries me a bit - possible
GetArgument with types that support only subset of range (but it looks like it
can happen even now for i32 vs i64).

Reviewed By: dzhulgakov

Differential Revision: D4611482

fbshipit-source-id: 77917fd5e1d18a1b860e022ede4518143d0f3f26