Add NanCheckOp, an operator that checks for NaNs and inf's on both the forward and backward pass.
Summary: NanCheck is an in-place operator for GPU that checks the input for any NaN or inf values. The operator fails and prints diagnostic information (input tensor dims and values) if it detects these erroneous values. This should help us to narrow down our numerical instability issues in the NMT models, and it might help others as well.

Differential Revision: D4818141

fbshipit-source-id: e5aa9762089c58ce160270446007c7a91a7a85e5