fix softmax ops dimension, add explicit rowmax buffer for simplicity
Summary: scale_ tensor was resizd in correctly in SoftmaxOp CUDA version. For some reason this has not triggered more crashes. I was using the rowmax_ in-place with scale_, which was then also incorrectly sized. Usually D>N, so this was not a issue, but perhaps there were cases with attention where this is not the case. Also the problem is order-sensitive, since if we once had an input with large D, the buffer was of correct size.

Reviewed By: jamesr66a

Differential Revision: D4904989

fbshipit-source-id: 244b6d308d1fc08be885c641440cbacad3b0dbce