better TileOp/Gradient CUDA implementation
Summary: ves and jamesr66a had noticed that TileOp for CUDA was very slow, as it started kernels inside double loops. It was my fault not to notice this in the code review. This diff uses 1 kernel for forward and backward passes and is probably much faster. I did not test though, maybe ves or jamesr66a can help?

Reviewed By: jamesr66a

Differential Revision: D5101968

fbshipit-source-id: 64b6ac933785e3710b3c1d8c692a4c48650bca96