fix for back-and-forth models, pass reference instead of copy
Summary:
akirillov again presented me with a memonger-bug: his model that has kind of a 'back-and-forth structure' where blobs are passed left and right in a ladder-like structure, revealed a bug in memonger: I should pass the set of free blobs as a reference, not a copy so that the recyclings are properly accounted for. Hard to explain.

Since we have the graph verifier, we can be more confident with these changes.

I also added some helpful debug to the graph verifier.

Differential Revision: D5396925

fbshipit-source-id: 0bffb3a0bf8532afcd6b5bc9331c779768a8c5c5