Use Toffee IR schema to disambiguate types of attributes.
Let say I write alpha=2 in my PyTorch code.  Is alpha a float
or an int?  This problem is resolved when we actually pass
it to the underlying kernel, which knows what type it expects
it as.

When serializing to Toffee IR, the Toffee NodeProto also needs
to dictate the correct type; otherwise, we may guess wrong.
We get this information from the OpSchema in the ToffeeIR library.
With this, we can avoid explicitly casting in dropout.py and
auto_primspec.py

WARNING: You will need to update torch/lib/ToffeeIR when you pull
this patch, as attribute schemas were added recently to ToffeeIR.

Signed-off-by: Edward Z. Yang <ezyang@fb.com>