add axis argument to NormalizeOp and NormalizeGradientOp
Summary:
As described in task T21337239, NormalizeOp currently normalizes over only the last dimension.
In this commit, the following changes have been made:
(1) Added an axis-parameter to NormalizeOp in both the CPU and CUDA context.
(2) Added the same axis parameter to  NormalizeGradient in both the CPU and CUDA context
(3) Removed the limit that the original NormalizeOp operator requires the input dimension to be 2

Reviewed By: akyrola

Differential Revision: D5745162

fbshipit-source-id: 69e04f59ac4d954b0062c3b2a53c8ca465a1027b