Add `@ignore` for script classes (#23614)
Summary:
This lets you mark a class so that it won't be recursively compiled.

This also runs up against a weird thing on the UX side, that to ignore a
module you have to `ignore` its `forward()` method but to ignore a
class you use `ignore` on the class declaration. The `ignore` on the
class declaration matches the use of `script` for script classes but is
confusing to those that don't know the difference between script classes
/ modules.
](https://our.intern.facebook.com/intern/diff/16770068/)
Pull Request resolved: https://github.com/pytorch/pytorch/pull/23614

Pulled By: driazati

Differential Revision: D16770068

fbshipit-source-id: bee9a9e88b6c798ce779f622c4f929adae4eaf45