add enabled API to autograd profiler (#31380)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/31380

For being able to profile async RPCs, we attach a `RecordFunction` object to the future that is created during the RPC to persist it across the lifetime of the RPC (this is implemented in the next PR: ). Since we'd only like to do this when profiling is enabled, this PR adds an enabled API to the autograd profiler.
ghstack-source-id: 96053933

Test Plan: Modified unit test.

Differential Revision: D19050391

fbshipit-source-id: aa382110e69d06b4a84c83b31d2bec2d8a81ba10