Remove ImplicitTensorToNum (#32761)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/32761

This replaces ImplicitTensorToNum with result-specific operators like
IntImplicit, FloatImplicit, or ScalarImplicit. Note that ScalarImplicit
was not correctly implemented before and this PR fixes the lapse.

This does not change on-disk serialization because these operators are not
serialized directly but written as eg. `annotated(int, foo)`.

Test Plan: Imported from OSS

Differential Revision: D19615385

Pulled By: zdevito

fbshipit-source-id: 48575f408e8219d2ec5b46936fc2aa691f283976