[JIT] Allow mutated values as functional graph inputs (#33297)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/33297

Allowing mutated values as inputs but not outputs has the effect of buffering up all mutated values as inputs to the graph. Just as we values which escape scope as graph inputs but not graph outputs - we should also allow values that get mutated. In both cases, the contract is that that the functional graph cannot write to graph inputs.

Without this patch, if there is a single write to the Tensor wildcard set it would disable all optimization.

Test Plan: Imported from OSS

Differential Revision: D20607175

Pulled By: eellison

fbshipit-source-id: c698e7cf3374e501cd5d835663991026a113ec6b