_requires_grad -> requires_grad (#36168)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/36168

This makes it match what eager mode uses as the keyword name.
Currently _requires_grad will not appear in serialization because
it is not listed as kwarg only. There is a small chance there is a model
that has never been run in eager mode that uses the _requires_grad name,
but this is rare enough that I don't think we need to worry about it unless
something breaks in testing.

Test Plan: Imported from OSS

Differential Revision: D20902557

Pulled By: zdevito

fbshipit-source-id: 605cf5371b4fc15ec1b4e8a12f9660d723530de4