Logging for ATen op subtype
Summary: ATenOp should go away, but before it does it's important to understand what's going inside of it. We already log `arguments`, but it's rather hard to parse in scuba as its a list, not a dictionary. Let's extract operator name explicitly so that grouping works well

Test Plan: unittest

Reviewed By: ngimel

Differential Revision: D21057966

fbshipit-source-id: 86be7cca39055620477a28bd5d8ab29e8edd2ff9