[caffe2] Fast path for single tensor in UnPackRecordsOp (#37361)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/37361

Add a fast path for the case of batch_size = 1 and single ad embedding in UnPackRecordsOp. In this case, there is no need to pack the single tensor into a shared_ptr<vector<vector<Tensor>>> and then unpack it in UnPackRecordsOp. Instead, we can just pass the tensor as it is into UnPackRecordsOp and share the data with the output tensor.

Reviewed By: yinghai

Differential Revision: D21224497

fbshipit-source-id: 70685e5cc20ffdc5e0044a4b97a7fc5133786db4