Fix `cdist` backward calculation for `p=2` (#37337)
Summary:
Closes https://github.com/pytorch/pytorch/issues/37154

Fixes a bug in `cdist` backward with `p=2`.
Under some circumstances, if the output has 0s, the gradient calculation of `sqrt` will be undefined. Leading to NaNs in the input gradients.

This PR defines a subgradient for this case.

A test is also added to verify this behavior, I was only able to reproduce it under certain shapes, so the shape is explicitly taken from https://github.com/pytorch/pytorch/issues/37154 example
Pull Request resolved: https://github.com/pytorch/pytorch/pull/37337

Differential Revision: D21403178

Pulled By: albanD

fbshipit-source-id: deef9678c1958524b552504920f19617f9ad1da6