Also set CMAKE_C_STANDARD for MSVC (#39304)
Summary:
According to
<https://gitlab.kitware.com/cmake/cmake/-/blob/master/Modules/Compiler/MSVC-C.cmake>,
the option simply has no effect for MSVC as of today. It is better to not impose
such an if condition as it is a bit misleading (the current code makes it look like we have compatibility issues with MSVC C11 support), and also it's better to
leave the judgment of MSVC C support to CMake devs.
Pull Request resolved: https://github.com/pytorch/pytorch/pull/39304

Differential Revision: D21846032

Pulled By: malfet

fbshipit-source-id: 962e5721da3d7b9be4117b42bdc35df426b7da7b