Fix FC issue (#41198)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/41198

https://github.com/pytorch/pytorch/pull/39611 unified signatures of some ops taking TensorOptions arguments by making them optional.
That has FC implications but only for models writting with a PyTorch version after that version (see explanation in description of that PR).

However, it also changed the default from `pin_memory=False` to `pin_memory=None`, which actually breaks FC for preexisting models too if they're re-exported with a newer PyTorch,
because we materialize default values when exporting. This is bad.

This PR reverts that particular part of https://github.com/pytorch/pytorch/pull/39611 to revert the FC breakage.
ghstack-source-id: 107475024

Test Plan: waitforsandcastle

Reviewed By: bhosmer

Differential Revision: D22461661

fbshipit-source-id: ba2776267c3bba97439df66ecb50be7c1971d20d