[Distributed] Set smaller Store timeouts to make c10d tests run faster (#46067)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/46067

In our store tests, we expect there to be an exception when we call
get on a recently-deleted key. Unforunately, the store waits for the timeout
period for the key to be set before throwing, which causes the tests to idel
wait for 5+ minutes. This PR decreases the timeouts before this set call so
these tests run faster.
ghstack-source-id: 113917315

Test Plan: Ran both the Python and C++ tests.

Reviewed By: pritamdamania87

Differential Revision: D24208617

fbshipit-source-id: c536e59ee305e0c01c44198a3b1a2247b8672af2