Fix multiple spelling and grammar mistakes (#48592)
Summary:
I found a number of spelling & grammatical mistakes in the repository. Previously I had these fixes submitted individually, but I saw that a single word change was apparently too small for a PR to be merged. Hopefully this new PR has a sufficient number of changes.

Pull Request resolved: https://github.com/pytorch/pytorch/pull/48592

Reviewed By: ejguan

Differential Revision: D25224216

Pulled By: mrshenli

fbshipit-source-id: 2af3db2aee486563efd0dffc4e8f777306a73e44