Add deadline to fakelowp tests (#48823)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/48823

deadline=None is not good because

Sandcastle tests will return success for tests timeout (default flag), and we cannot efficiently detect broken tests if there is any.

In addition, the return signal for timeout is 64, which is same as skip test.

Test Plan: Sandcastle, and run tests on card

Reviewed By: hyuen

Differential Revision: D25318184

fbshipit-source-id: de1b55a259edb2452fb51ba4c598ab8cca9e76b7