[PyTorch] Add VariableVersion&& overload for TensorImpl::shallow_copy_and_detach (#48681)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/48681

This should reduce reference counting traffic when creating views.

The code duplication here is unfortunate and I'm open to suggestions on how to reduce it. It's especially regrettable that we create a footgun for subclasses of TensorImpl: they can accidentally override only one of the two overloads and get confusing behavior.
ghstack-source-id: 117896685

Test Plan: internal benchmarks

Reviewed By: ezyang

Differential Revision: D25259741

fbshipit-source-id: 55f99b16b50f9791fdab85cbc81d7cd14e31c4cf