Add some safeguards to FutureNCCL (#48562)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/48562

This commit is part of a stack that reworks FutureNCCL in order to extract a generic CUDA-aware Future subclass. The stack deliberately breaks up this transition into elementary changes, to make it easier to verify that the behavior is preserved (or to highlight how it gets changed).

 ---

In this commit I'm adding a few asserts to the constructors of FutureNCCL to make sure that what's passed in is what we expect (fun fact: until two commits ago that wasn't the case, as we were passed some empty events).

I'm also making the second constructor private, as it's only supposed to be used by the then() method.
ghstack-source-id: 118180036

Test Plan: Unit tests

Reviewed By: mrshenli

Differential Revision: D25210333

fbshipit-source-id: d2eacf0f7de5cc763e3cdd1ae5fd521fd2eec317