[JIT] disable masked fill (#50147)
Summary:
There is an internal user who is experiencing a bug with masked_fill. While I am almost certain this corresponds to an old pytorch version with the bug, the model that is breaking is important and time-sensitive and we are covering all bases to try to get it to work again.

Pull Request resolved: https://github.com/pytorch/pytorch/pull/50147

Reviewed By: nhsoukai

Differential Revision: D25806541

Pulled By: eellison

fbshipit-source-id: 131bd71b5db9717a8a9cb97973d0b4f0e96455d6