Add ARM64 cross-compilation build on OS X (#49751)
Summary:
Tests cross-compilation of ARM64 architecture in MacOS CI.

This should be merged after PR https://github.com/pytorch/pytorch/issues/50243 and https://github.com/pytorch/pytorch/issues/50922 (adding a fix).

The reason we pin the wheel to be version 0.36.2 is because lower versions cannot handle c38 as a tag for the wheel.

Pull Request resolved: https://github.com/pytorch/pytorch/pull/49751

Reviewed By: albanD

Differential Revision: D26411133

Pulled By: janeyx99

fbshipit-source-id: 00a5cf597aee10adea1547579270cb3b38732563