[caffe2] add a CAFFE2_NODISCARD macro to help support old compilers (#53754)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/53754

Some of the PyTorch CircleCI builds still use gcc 5.4, and compile with
`-Werror=attributes` causing this old compiler to fail because it does not
understand the `[[nodiscard]]` attribute.

Let's define a `CAFFE2_NODISCARD` macro to work around this.
ghstack-source-id: 123594084

Test Plan: I'm using this macro in subsequent diffs in the stack.

Reviewed By: mraway

Differential Revision: D26959584

fbshipit-source-id: c7ba94f7ea944b6340e9fe20949ba41931e11d41