[caffe2] Refactor tensor serialization function (#53404)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/53404

This refactors `TensorSerializer::Serialize()` so that we have a separate
helper function for each data type.

This should make it slightly easier in the future to add new serialization
formats for specific data types.
ghstack-source-id: 124085413

Test Plan:
Confirmed the existing tests pass.  This diff is not expected to have any
behavior changes.

Reviewed By: mraway, glamtechie

Differential Revision: D26658204

fbshipit-source-id: 232776262db6486ba845a7ba223e3987053dac27