Pin autocanceling GHA repo to specific commit (#54738)
Summary:
This way, if malicious code gets committed and the tag moves forward, we would be at risk. This does mean that we would have to manually update the SHA if there are desirable upgrades to the repository.

We are pinning it to this commit: https://github.com/potiuk/cancel-workflow-runs/commit/a81b3c4d59c61e27484cfacdc13897dd908419c9

Pull Request resolved: https://github.com/pytorch/pytorch/pull/54738

Reviewed By: samestep

Differential Revision: D27346792

Pulled By: janeyx99

fbshipit-source-id: 5641a78567c3cd61dce35dfa2fd4918f255a7681