Add breakpad to Docker image (#55439)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/55439

This adds [breakpad](https://github.com/google/breakpad) to the build in CI (just on one image for now). I attempted in #54739 to build it from source as a normal third_party submodule but it uses autotools and has some weird build steps that made it hacky to integrate. We really only need it for release builds anyways since its use is moot if built with anything but `RELEASE=1`.

Test Plan: Imported from OSS

Reviewed By: nikithamalgifb

Differential Revision: D27679766

Pulled By: driazati

fbshipit-source-id: 8211444df49b219c722137b9243d16d649a1f1ae