Adding super calls to JIT test case setUp and tearDown (#61922)
Summary:
This issue was surfaced when adding this issue: https://github.com/pytorch/pytorch/issues/61655 did not manage to skip the appropriate test case.

I then investigated and realized it was because the setUp code that does the test disabling is not called because another defined setUp overrode the parent class' setUp.

I am not sure if that was intentional--if so we would have to adopt the child class' code to call the check_if_enable function in common_utils.

Pull Request resolved: https://github.com/pytorch/pytorch/pull/61922

Reviewed By: ejguan

Differential Revision: D29798716

Pulled By: janeyx99

fbshipit-source-id: d31b664e48507d69de14574ff5e6ecf1d41ae24d