Add build/.ninja_log to artifacts for Windows (#62035)
Summary:
Being able to download the .ninja_log allows for better debugging. There may be a follow-up PR to convert this to a better tracefile.

This PR only handles windows as it is already handled for linux here:
https://github.com/pytorch/pytorch/blob/master/.jenkins/pytorch/build.sh#L248-L252

Pull Request resolved: https://github.com/pytorch/pytorch/pull/62035

Test Plan: Check the artifacts for a windows job and see if we see .ninja_log

Reviewed By: malfet

Differential Revision: D29852228

Pulled By: janeyx99

fbshipit-source-id: a3a87b709cd0c84f5b3cdc274ac4a623771c2b5c