Deprecate x.T on tensors of dimension other than 0 or 2 (#64180)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/64180

**BC-breaking note:**

This PR deprecates the `Tensor.T` are not matrices. An upgrade guide is added to the
documentation for `Tensor.T`.

This PR DOES NOT make this attribute to throw an error when called on a tensor of `dim != 2`,
but this will be its behavior in a future PyTorch release.

cc mruberry rgommers pmeier asmeurer leofang AnirudhDagar asi1024 emcastillo kmaehashi heitorschueroff

Test Plan: Imported from OSS

Reviewed By: bdhirsh

Differential Revision: D31610611

Pulled By: anjali411

fbshipit-source-id: af8ff7e862790dda9f06921de005b3f6fd0803c3