[SR] Fix bug in FuseListUnpackV2 (#67021)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/67021

When applying the equally split optimization, we still need to delete the list unpack node.

I did an accuracy test yesterday but didn't catch this issue because my diffs were not properly synced between devservers (I use hlu1's devbig for testing and it had an old version of "Add FuseListUnpackV2"). But I did another test this morning and realized that there was an issue.

This is not affecting anything in prod right now since D31742293 has not landed.

Reviewed By: hlu1

Differential Revision: D31827278

fbshipit-source-id: c7b05e3d8ec942632adcff4bdfebb8c27c1a7a39