[forward ad] Also check layout of grad matches that of self for inplace over view (#67816)
Summary:
Fixes https://github.com/pytorch/pytorch/issues/67800

Currently when the grad is the same layout as base, we try to assign the same tensor to the forward grad of both the base and the view. However, when the layout of the grad is different from the layout of the view, this triggers a copy to be created, and the tangent of the view (after the inplace) will not have a view relationship with the view of the base.

This PR just changes it so that we only do the above optimization when the layout also matches the layout of self

Pull Request resolved: https://github.com/pytorch/pytorch/pull/67816

Reviewed By: malfet

Differential Revision: D32190021

Pulled By: soulitzer

fbshipit-source-id: b1b2c9b332e83f4df5695ee9686ea76447f9305b