Fixes deprecation warnings in `test_optim.py` (#67954)
Summary:
Catches deprecation warnings when we call `scheduler.step(epoch)`
in tests.

Removes duplicate parameters to optimizers unless we are specifically
testing for that

Fixes https://github.com/pytorch/pytorch/issues/67696

There is one warning remaining when I run this locally -- however that is due to the implementation of the `SequentialLR` Scheduler. I will open a new issue relating to that.

Pull Request resolved: https://github.com/pytorch/pytorch/pull/67954

Reviewed By: H-Huang

Differential Revision: D32244056

Pulled By: albanD

fbshipit-source-id: 2ab3086a58e10c8d29809ccbaab80606a1ec61d8