[PyTorch Edge] Better error message when training attribute is not found (#68103)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/68103

The error message `'training' attribute not found.` in itself isn't particularly actionable. Anyone running into this tends to be clueless regarding why they are getting this message.

For example, see [this post](https://fb.workplace.com/groups/pytorch.edge.users/posts/965868874283406/) asking for help when seeing this specific error message.

The most common reason for this error is that users call `.eval()` in the model instance before saving it. This change tries to draw attention to that oversight and allows them to proactively investigate and correct that mis-action if necessary.

This saves valuable time for our users and effort from the team tp provide support. Overall, I believe this is a Developer Experience win.

ghstack-source-id: 143021300

Test Plan: Build/CI

Reviewed By: JacobSzwejbka

Differential Revision: D32304477

fbshipit-source-id: 474abe717a862347f16ad981834ddab6819cb4d3