dbr quant: fix debugging fqn info for converted model (#68840)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/68840

Fixes the debugging FQN info for a converted model. Some of this
information was missing because eager mode convert performed
module swaps. This information is only used in debugging and is
not used for inference.

Test Plan:
```
python test/test_quantization.py TestQuantizeDBR
```

turn `enable_logging` on in `auto_trace.py`, the FQN is now displayed
for a converted model

Reviewed By: jerryzh168

Differential Revision: D32630884

Pulled By: vkuzo

fbshipit-source-id: be8c43343abfdab9fe0af39499d908ed61a01b78