Fix test tools
This is a temp fix to get trunk green (uses newer stats). The reason the test failed before was because the tests referred to old stats that were deleted after their lifecycle in S3.

However, we need to come up with a better way of testing this as this same failure will occur again in a few months.
Pull Request resolved: https://github.com/pytorch/pytorch/pull/73082