[mergebot] Make merge on green default behavior (#79199)
This makes the merge on green behavior default as we tried to do it earlier, but forgot to pass in the force arguments.

This also adds a few tests so that we don't make that mistake again (although people can still forget to pass in the arguments through shell).
Pull Request resolved: https://github.com/pytorch/pytorch/pull/79199
Approved by: https://github.com/janeyx99, https://github.com/seemethere, https://github.com/malfet