fix up trymerge (#80100)
Addresses #80096

In the Github API, they don't set the parent conclusion to FAILURE if one of the jobs has a failure. It's stuck at null until all jobs finish, which isn't good for TTS on mergebot.

Example GH response:
<img width="827" alt="image" src="https://user-images.githubusercontent.com/34172846/175170084-de6bb419-925e-41a6-8c12-6732b35b6cab.png">
<img width="1264" alt="image" src="https://user-images.githubusercontent.com/34172846/175170180-c629ed4d-a174-4179-8321-1a192d1ec8f3.png">

In here, we check if there's a failing check within the checkruns and if it is, it'll set the workflow run conclusion to failure

Test Plan:

Run the find_matching_merge_rule method and check that it throws a RunTimeError for pull on a certain PR

<img width="1172" alt="image" src="https://user-images.githubusercontent.com/34172846/175170413-a28352c6-f518-4aa2-a068-d13df9f76268.png">

Pull Request resolved: https://github.com/pytorch/pytorch/pull/80100
Approved by: https://github.com/janeyx99, https://github.com/seemethere