[mergebot] combine land check and job started comments (#80965)
Addresses https://github.com/pytorch/pytorch/issues/80923

The alternative to this is editing the start comment to include the land checks details, but I think it might be a bit misleading because the land check branch creation might not be there when the person first clicks on it when they get a notification.  This might be a bit annoying cause then the user won't be able to look at their progress.
Pull Request resolved: https://github.com/pytorch/pytorch/pull/80965
Approved by: https://github.com/malfet