[mergebot] Add Land Check Troubleshooting Message (#82580)
### Description
Land checks are new and people can get confused on how to use it. In this, we make it a bit clearer to the user that they can use pytorchbot merge -g or -f to merge it in if they believe that there's infra issues with the land check.

### Issue
N/a

### Testing
N/a. Lint should be enough.
Pull Request resolved: https://github.com/pytorch/pytorch/pull/82580
Approved by: https://github.com/huydhn