[docs] fix inconsistent default `rcond` value (#82887)
The default `rcond` value is described as "the machine precision of the dtype of :attr:`A`" in the text (line 1043) but "the machine precision of the dtype of :attr:`A` times `max(m, n)`" in the `Args` section (line 1079). The correct value, according to [this issue](https://github.com/pytorch/pytorch/issues/82868) is ":attr:`A` times `max(m, n)`", so I'm updating the description to match.

### Description
<!-- What did you change and why was it needed? -->

### Issue
https://github.com/pytorch/pytorch/issues/82868

### Testing
<!-- How did you test your change? -->

Pull Request resolved: https://github.com/pytorch/pytorch/pull/82887
Approved by: https://github.com/IvanYashchuk