Temporary fix to not fail concurrent viable/strict updates (#84324)
Until we have a solution for https://github.com/pytorch/pytorch/issues/83986 and can use our runner for the job, we need to live with the fact that GitHub runner can have a pretty long queue throughout the day.  This keeps trunk green in the meantime.

This's a follow-up of https://github.com/pytorch/pytorch/pull/84249
Pull Request resolved: https://github.com/pytorch/pytorch/pull/84324
Approved by: https://github.com/zengk95