Remove setup-python on Windows CI and use Conda instead (#92183)
This has been bugging me for a while what Windows CI stills has the `setup-python` step in its setup.  The python setup here is not used by the build and the test steps at all, but are there to provide a python3 interpreter for `actions/get-workflow-job-id` and `actions/filter-test-configs`.  ~~As these 2 actions are generic and should be smart enough to check for conda setup and use that instead of system python.~~

Having `setup-python` contributes a bit to network flakiness on Windows where it fails to download stuffs from GitHub.  Example failures:

* https://github.com/pytorch/pytorch/actions/runs/3913257969/jobs/6690485582
* https://github.com/pytorch/pytorch/actions/runs/3930859163/jobs/6722743854
* https://github.com/pytorch/pytorch/actions/runs/3918415654/jobs/6699239557
Pull Request resolved: https://github.com/pytorch/pytorch/pull/92183
Approved by: https://github.com/ZainRizvi